backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
tiktoken
beautifulsoup4
pdfminer.six
youtube-transcript-api
requests
wikipedia
pydantic
pydantic-settings
prisma
passlib[bcrypt]
pyjwt
groq
langchain-groq
